test(io): dedup box-row/prism fixtures and progress recorder (#1281, #1282)
Pass 5d tests-duplication cluster: two findings that share files with each
other, fixed together.

#1281: the "row of N translated 10x10x10 boxes" and "1200-sided N-gon prism,
r=1000, h=50" fixtures were each reimplemented independently across the
robust-import cancellation suites, differing only in loop count and (for the
prism) the STEP-write step. Added two shared factory functions to
IOTestFixtures.swift, `boxRow(count:)` and `ngonPrism(sides:radius:height:)`,
and replaced all five call sites (MultibodyRobustImportTests.tenBoxes(),
CancellationReportingTests.igesRobustTransferPhaseCancellationIsCancelled and
prismSTEP(named:), RobustImportProgressTests.igesRobustHealCancellation and
stepRobustRepairCancellation).

#1282: MeshAndExportProgressTests.Recorder and
ImportProgressTests.ProgressRecorder were the same lock-guarded
ImportProgress recorder under two names, Recorder exposing only
`eventCount: Int`, a strict subset of ProgressRecorder's `events`. Moved
ProgressRecorder to IOTestFixtures.swift as the one shared recorder and
switched MeshAndExportProgressTests' call sites to `events.count`.

No behaviour change: same geometry, same recorder semantics, only the
duplication removed. Verified with a focused `swift build --target
OCCTIOTests` and `swift test --filter` across all five affected suites (26
tests, 5 suites, all passing), plus a full `swift build`.

Closes #1281
Closes #1282
